pyproject.toml
[tool.poetry]
name = "commercialnetwork"
version = "0.1.0"
description = ""
authors = ["Maxim Abavi <maxim.abavi@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
psycopg = "^3.1.19"
django-phonenumber-field = "^7.3.0"
phonenumbers = "^8.13.37"
python-dotenv = "^1.0.1"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
setuptools = "^70.0.0"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
